Added ReactionNetwork and parse<T>().
Fixed #8 by adding parse<T>() in parse.h so that we have streamlined
parsing for structured blocks. Added ReactionNetwork and an initial test
for it, but not the flux calculation. Merged INIPropertyParser.cpp into
the header to do templating better. Updated documentation to reflect all
of these changes and various tests.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>