Got an initial solve working for the astro network.
Fixed the solvers package so that it can be linked from other modules.
Added initial SUNDIALS and CVODE support. Added a CVODE example from
their distribution to test integration in our build. Added an ODESolver
stub. Added what will be an integration test for the reaction network
with an initial CVODE solver crammed into the test. This will all be
refactored soon.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>